# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KETANTECH VPN Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, paid, failed, expired]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Retired static product purchase endpoint
      deprecated: true
      description: Static product purchases have been retired. Use POST /dynamic-vpn/orders for new purchases.
      security:
        - cookieAuth: []
      responses:
        "410":
          description: Static purchases are retired; use POST /dynamic-vpn/orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/pay:
    post:
      operationId: payOrder
      tags: [orders]
      summary: Retired static order payment endpoint
      deprecated: true
      description: Static order payments are retired. Use GET /orders for history and POST /dynamic-vpn/orders for new purchases.
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "410":
          description: Static order payments are retired; use order history or dynamic ordering.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Balance / Topup ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        replacement:
          type: string
...
    Product:
      type: object
      required: [id, name, protocol, durationDays, price, isActive, stock, availableStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        protocol:
          type: string
          enum: [ssh, vmess, vless, trojan, shadowsocks]
        durationDays:
          type: integer
        price:
          type: number
        quota:
          type: number
          nullable: true
          description: Bandwidth quota in GB, null = unlimited
        maxConnections:
          type: integer
          nullable: true
        stock:
          type: integer
          description: Batas maksimal akun aktif untuk produk ini
        availableStock:
          type: integer
          description: Sisa stok yang masih tersedia (stock - akun aktif)
        resellerPrice:
          type: number
          nullable: true
          description: Harga khusus reseller (null jika user bukan reseller atau fitur reseller nonaktif)
        isActive:
          type: boolean
        category:
          type: string
          nullable: true
        sortOrder:
          type: integer
        serverId:
          type: integer
          nullable: true
          description: ID server yang di-pin untuk produk ini (null = pilih otomatis)
        serverName:
          type: string
          nullable: true
          description: Nama server yang di-pin (null jika tidak di-pin)

    # ─── Orders ───────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, userId, productId, status, amount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
          nullable: true
        status:
          type: string
          enum: [pending, processing, paid, failed, expired]
          description: processing berarti pembayaran sudah diterima dan akun VPN sedang dibuat
        amount:
          type: number
          description: Harga dasar order dalam IDR, sebelum kode unik
        vpnAccountId:
          type: integer
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        paymentProvider:
          type: string
          nullable: true
          enum: [ketantechpay, autogopay]
          description: Penyedia pembayaran yang memproses transaksi; null untuk pembayaran saldo atau QRIS statis legacy
        paymentChannel:
          type: string
          nullable: true
          enum: [ketantechpay, autogopay_gopay, autogopay_shopeepay]
          description: "Channel pembayaran: ketantechpay = QRIS dinamis KetantechPay; autogopay_gopay = QRIS dinamis melalui channel GoPay AutoGoPay; autogopay_shopeepay = QRIS melalui channel ShopeePay AutoGoPay. QRIS interoperabel dapat dipindai dengan GoPay, OVO, ShopeePay, atau aplikasi QRIS lain; OVO bukan channel yang dipilih secara terpisah."
        payableAmount:
          type: number
          nullable: true
          description: Jumlah aktual dalam IDR yang harus dibayar, termasuk kode unik jika ada; gunakan amount bila null
        uniqueCode:
          type: integer
          nullable: true
          minimum: 0
          description: Tambahan kode unik pada payableAmount; setelah pembayaran berhasil nilai kode unik dikreditkan ke saldo pengguna
        notes:
          type: string
          nullable: true
        qrisUrl:
          type: string
          nullable: true
          description: URL gambar QRIS untuk paymentMethod=qris
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Waktu kedaluwarsa pembayaran QRIS
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
...